Fix 'Show all' facets being selected breaking the search
