+++ src/algolia/utils.js
@@ -8,3 +8,36 @@ export const facetPropPathMapping = {
   'attributes.subject.ageCategory.value' : 'Age Categories',
 }
 
+/* Returns filter for searching algolia. All facets of the same category are joined with OR,
+  * and each of those results is then joined with an AND.
+  * i.e. (color:blue OR color:red) AND (shape:circle OR shape:red) */
+export function getFilters(selectedFacetArray) {
+  if (selectedFacetArray === undefined) {
+    return undefined;
+  }
+  let facets = removeShowAllFacets(selectedFacetArray)
+
+  let filters = "NOT item.published.status:embargo";
+  filters = `(${filters}) AND `;
+
+  const facetPropPaths = Object.keys(facetPropPathMapping);
+  facetPropPaths.map((facetPropPath) => {
+    const facetsToOr = facets.filter(
+      (facet) => facet.facetPropPath == facetPropPath
+    );
+    var filter = "";
+    facetsToOr.map((facet) => {
+      filter += `"${facetPropPath}":"${facet.label}" OR `;
+    });
+    if (filter == "") {
+      return;
+    }
+    filter = `(${filter.substring(0, filter.lastIndexOf(" OR "))})`;
+    filters += `${filter} AND `;
+  });
+  return filters.substring(0, filters.lastIndexOf(" AND "));
+}
+
+function removeShowAllFacets(facetArray){
+  return facetArray.filter( f => f.label !== 'Show all')
+}
